[composables] ARCH-22 (1D.7): extract shared player transport state
PreviewPlayer, PlaylistPlayer and SharedPlaylistPlayer each declared their
own currentTimeRaw/isHd/isMuted/isPlaying/isRepeating/speed/volume refs with
identical defaults, one more place these three files had to be kept in sync
by hand for state that has nothing player-specific about it.

Add usePlayerTransport() in src/composables/players/transport.js, owning
only those refs, and destructure it at each player's State block in place
of the individual declarations. SharedPlaylistPlayer's isHd defaults to
true (not false) and its volume defaults to 100 (not 50) — both differ from
the shared defaults, so they stay as local refs and are excluded from that
file's destructure; SharedPlaylistPlayer also has no currentTimeRaw/speed
refs at all, so its destructure only pulls isMuted/isPlaying/isRepeating.
Side effects (preference persistence, decoder calls, room sync) are
untouched — this only moves declarations.
